ファイル無いとエラーになるらしいので、削除前に存在チェックするようにする

+++ tests/obsplan/test_mapsim1.py
@@ -4,7 +4,8 @@ import filecmp
 import os
 
 def test_it():
-    os.remove("map.svg")
+    if os.path.isfile("map.svg"):
+            os.remove("map.svg")
     make_image()
     assert(filecmp.cmp("map.svg", "../tests/obsplan/out.svg"))
     os.remove("map.svg")
